move all env vars to environment package
